+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Add event not found handling
 - Fix dynamic avatar layer removability at extra small
 - Fix dynamic avatar image paths
+- Fix challenge task positioning
 
 ## [4.0.0-alpha5] 2017-05-01
 ### Added